+++ CHANGELOG.md
@@ -4,7 +4,7 @@ This change log follows the conventions of [keepachangelog.com](http://keepachan
 ## [Unreleased]
 ### New
 - Added “maybe” wrappers for `rotate`, `scale` and `translate`.
-- Added threaded fasteners.
+- Added threaded fasteners and a script showcasing a selection of these.
 - Added `long-hex-diagonal` for computing the long diagonal of a hexagon
   from the short diagonal.
 - Added a handful of constants and abbreviations of dubious value: